refactor: fix Directionable to Directable
